NIFI-11259 - ConsumeKafka commit offsets test refactor
